RavenDB-17771 cleanup
